Setup Jenkins Docker client using user's Triton environment
Use the user's environment variables to populate the minimum number of variables
we need to run sdc-docker-setup.sh in the Jenkins container. The setup.sh script
will write these to an _env file and the onStart handler will use them to
bootstrap from there.

As written, this won't *quite* work when we launch a Jenkins container locally
for testing, so it needs some adjustment.