Jenkins job to create Jenkins jobs from GitHub, with example
The intended design here is for Jenkins jobs to be stored under source control
(and we're assuming git here). Added `get_jobs` to the setup script to inject
the env vars into a job called "jenkins-jobs" that we'll install at startup.
This job itself will query the jobs repo and copy the jobs from that repo into
the appropriate location so that Jenkins can run them.

This commit includes an example job in this repo so that we can have this repo
act as its own jobs repo for demonstration purposes.

Write our Jenkins job GitHub repo/refspec into a template job that