Add check of job-builder job to health check
When we run the onStart handler Jenkins isn't running yet, so we can't run the
first job-building job to get the rest of the jobs from Jenkins. So we add a
filesystem check to see if the job-building job has ever been run, and if not
we fire a REST API call to Jenkins to force it to run once. At this point, we
assume GitHub webhooks, polling, scheduled jobs, etc. will take over.